+++ AddOns/Baganator/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Baganator
 
-## [511](https://github.com/Baganator/Baganator/tree/511) (2024-11-27)
-[Full Changelog](https://github.com/Baganator/Baganator/compare/510...511) 
+## [514](https://github.com/Baganator/Baganator/tree/514) (2024-11-28)
+[Full Changelog](https://github.com/Baganator/Baganator/compare/513...514) 
 
-- Fix dropping item via drag only (no clicks to pickup) for an item onto blank space  
+- Categories: Fix items skipping recents, properly this time.